8.2.3. Main page - setFavoriteStatus
