+++ project/src/components/app/app.tsx
@@ -1,3 +1,4 @@
+import {useEffect} from 'react';
 import {Route, BrowserRouter, Routes} from 'react-router-dom';
 
 import {PrivateRoute} from '../../components/private-route';
@@ -12,13 +13,20 @@ import {
   NotFoundScreen,
 } from '../../pages';
 
-import {fetchOffersAction} from '../../store/api-actions';
+import {fetchOffersAction, fetchFavoritesAction} from '../../store/api-actions';
+import {setFavoriteStatus} from '../../store/main-process/main-process';
 import {
   getActiveCity,
   getOffers,
   getIsOffersLoading,
   getIsOffersLoaded
 } from '../../store/main-process/selectors';
+
+import {
+  getFavorites,
+  getIsFavoriteLoading,
+  getIsFavoritesLoaded
+} from '../../store/favorite-data/selectors';
 
 import withMap from '../../hocs/with-map';
 
@@ -30,12 +38,38 @@ const MainScreenWrapped = withMap(MainScreen);
 const RoomScreenWrapped = withMap(RoomScreen);
 
 function App(): JSX.Element {
+  const {
+    Main,
+    MainEmpty,
+    Favorites,
+    FavoritesEmpty,
+    OfferId,
+    Login
+  } = AppRoute;
+
   const dispatch = useAppDispatch();
 
   const activeCity = useAppSelector(getActiveCity);
   const offers = useAppSelector(getOffers);
   const isOffersLoading = useAppSelector(getIsOffersLoading);
   const isOffersLoaded = useAppSelector(getIsOffersLoaded);
+
+  const favorites = useAppSelector(getFavorites);
+  const isFavoritesLoading = useAppSelector(getIsFavoriteLoading);
+  const isFavoritesLoaded = useAppSelector(getIsFavoritesLoaded);
+
+  useEffect((): void => {
+    if (!isFavoritesLoaded && !isFavoritesLoading) {
+      dispatch(fetchFavoritesAction());
+      <Loader />;
+    }
+  },[]);
+
+  useEffect((): void => {
+    if (favorites.length && isFavoritesLoaded) {
+      dispatch(setFavoriteStatus(favorites));
+    }
+  },[favorites]);
 
   if (!offers.length && !isOffersLoading && !isOffersLoaded) {
     dispatch(fetchOffersAction(activeCity));
@@ -46,15 +80,6 @@ function App(): JSX.Element {
       <Loader />
     );
   }
-
-  const {
-    Main,
-    MainEmpty,
-    Favorites,
-    FavoritesEmpty,
-    OfferId,
-    Login
-  } = AppRoute;
 
   return (
     <BrowserRouter>
