+++ project/src/components/favorites-screen/favorites-list.tsx
@@ -1,32 +1,14 @@
-import {useEffect} from 'react';
-
 import FavoritesEmptyScreen from '../../pages/favorites-empty-screen/favorites-empty-screen';
 import {FavoritesCard} from './';
-import {Loader} from '../../components/common';
 
-import {fetchFavoritesAction} from '../../store/api-actions';
-import {
-  getFavorites,
-  getIsFavoriteLoading,
-  getIsFavoritesLoaded
-} from '../../store/favorite-data/selectors';
+import {getFavorites, getIsFavoritesLoaded} from '../../store/favorite-data/selectors';
 
-import {useAppDispatch, useAppSelector} from '../../hooks';
+import {useAppSelector} from '../../hooks';
 
 /* «Список избранных предложений» */
 function FavoritesList(): JSX.Element {
-  const dispatch = useAppDispatch();
-
   const favorites = useAppSelector(getFavorites);
   const isFavoritesLoaded = useAppSelector(getIsFavoritesLoaded);
-  const isFavoritesLoading = useAppSelector(getIsFavoriteLoading);
-
-  useEffect((): void => {
-    if (!isFavoritesLoaded && !isFavoritesLoading) {
-      dispatch(fetchFavoritesAction());
-      <Loader />;
-    }
-  },[]);
 
   return (
     <div className="page">
