+++ project/src/utils.ts
@@ -16,3 +16,6 @@ export const sortTo = (array: any[], key: string, direction?: string): any[] =>
     /* eslint-disable-next-line */
     return direction === Order.Asc ? (a < b ? -1 : a > b ? 1 : 0) : (a < b ? 1 : a > b ? -1 : 0);
   });
+
+export const arrayToMap = (arr:any, byKey: string) =>
+  arr.reduce((acc: any, item: any) => ({...acc, [item[byKey]]: item}), {});
